requirements.txt
python-dotenv==1.0.0
langchain==0.2.16
langchain-community==0.2.16
langchain-postgres==0.0.9
langchain-huggingface==0.0.3
psycopg[binary]==3.2.1
psycopg2-binary==2.9.9
sentence-transformers==2.7.0
torch==2.3.1
transformers==4.42.4
numpy==1.24.4
